fix(Cache) apply corebos cache in listview and ispermitted method

+++ include/ListView/ListViewController.php
@@ -595,28 +595,44 @@ class ListViewController {
 
 			//Added for Actions ie., edit and delete links in listview
 			$actionLinkInfo = '';
-			if (isPermitted($module, 'EditView', $recordId) == 'yes') {
-				$racbr = $wfs->getRACRuleForRecord($currentModule, $recordId);
-				if (!$racbr || $racbr->hasListViewPermissionTo('edit')) {
-					$edit_link = $this->getListViewEditLink($module, $recordId);
-					if (isset($navigationInfo['start']) && $navigationInfo['start'] > 1 && $module != 'Emails') {
-						$actionLinkInfo .= "<a href=\"$edit_link&start=".$navigationInfo['start']."\">".getTranslatedString('LNK_EDIT', $module).'</a> ';
-					} else {
-						$actionLinkInfo .= "<a href=\"$edit_link\">".getTranslatedString('LNK_EDIT', $module).'</a> ';
+			$cacheId = $module.'#EditViewActions#'.$recordId.'#'.$current_user->id;
+			if ($cache->isUsable() && $cache->getCacheClient()->has($cacheId)) {
+				$actionLinkInfo .= $cache->getCacheClient()->get($cacheId);
+			} else {
+				if (isPermitted($module, 'EditView', $recordId) == 'yes') {
+					$racbr = $wfs->getRACRuleForRecord($currentModule, $recordId);
+					if (!$racbr || $racbr->hasListViewPermissionTo('edit')) {
+						$edit_link = $this->getListViewEditLink($module, $recordId);
+						if (isset($navigationInfo['start']) && $navigationInfo['start'] > 1 && $module != 'Emails') {
+							$actionLinkInfo .= "<a href=\"$edit_link&start=".$navigationInfo['start']."\">".getTranslatedString('LNK_EDIT', $module).'</a> ';
+						} else {
+							$actionLinkInfo .= "<a href=\"$edit_link\">".getTranslatedString('LNK_EDIT', $module).'</a> ';
+						}
+					}
+					if ($cache->isUsable()) {
+						$cache->getCacheClient()->set($cacheId, $actionLinkInfo);
 					}
 				}
 			}
 
-			if (isPermitted($module, 'Delete', $recordId) == 'yes') {
-				$racbr = $wfs->getRACRuleForRecord($currentModule, $recordId);
-				if (!$racbr || $racbr->hasListViewPermissionTo('delete')) {
-					$del_link = $this->getListViewDeleteLink($module, $recordId);
-					if ($actionLinkInfo != '' && $del_link != '') {
-						$actionLinkInfo .= ' | ';
+			$cacheId = $module.'#DeleteActions#'.$recordId.'#'.$current_user->id;
+			if ($cache->isUsable() && $cache->getCacheClient()->has($cacheId)) {
+				$actionLinkInfo .= $cache->getCacheClient()->get($cacheId);
+			} else {			
+				if (isPermitted($module, 'Delete', $recordId) == 'yes') {
+					$racbr = $wfs->getRACRuleForRecord($currentModule, $recordId);
+					if (!$racbr || $racbr->hasListViewPermissionTo('delete')) {
+						$del_link = $this->getListViewDeleteLink($module, $recordId);
+						if ($actionLinkInfo != '' && $del_link != '') {
+							$actionLinkInfo .= ' | ';
+						}
+						if ($del_link != '') {
+							$actionLinkInfo.="<a href='javascript:confirmdelete(\"".addslashes(urlencode($del_link))."\")'>".getTranslatedString('LNK_DELETE', $module).'</a>';
+						}
 					}
-					if ($del_link != '') {
-						$actionLinkInfo.="<a href='javascript:confirmdelete(\"".addslashes(urlencode($del_link))."\")'>".getTranslatedString('LNK_DELETE', $module).'</a>';
-					}
+				}
+				if ($cache->isUsable()) {
+					$cache->getCacheClient()->set($cacheId, $actionLinkInfo);
 				}
 			}
